refactor: CRUD api 파일 네이밍 수정(더 명확한 이름)
